Extract before_request decorator
