+++ cg/server/endpoints/sample_endpoints.py
@@ -4,11 +4,14 @@ from flask import Blueprint, jsonify, request
 
 from cg.server.dto.samples.collaborator_samples_request import CollaboratorSamplesRequest
 from cg.server.dto.samples.samples_response import SamplesResponse
+from cg.server.endpoints.utils import before_request
 from cg.server.ext import sample_service
 
 
 LOG = logging.getLogger(__name__)
 SAMPLES_BLUEPRINT = Blueprint("samples", __name__, url_prefix="/api/v1")
+
+SAMPLES_BLUEPRINT.before_request(before_request)
 
 
 @SAMPLES_BLUEPRINT.route("/samples_in_collaboration")
